+++ src/types/BlockFromPegasus.ts
@@ -1,4 +1,4 @@
-import { HexStringWith0x } from './HexStringWith0x';
+import { HexStringWith0x } from './custom';
 
 export interface BlockFromPegasus {
   _id: string;
@@ -7,8 +7,8 @@ export interface BlockFromPegasus {
   blockId: number;
   root: string;
   data: Record<string, HexStringWith0x>;
-  numericFcdKeys: Array<string>;
-  numericFcdValues: Array<number>;
+  fcdKeys: Array<string>;
+  fcdValues: Array<number>;
   anchor: string;
   votes: Record<string, string>;
   power: string;